working: folder nav, file open
